Nessus_Scan_Rule.sh
This script uses the Nessus and Defender command-line utilities to fetch lists of devices that have been scanned by each system. It then compares the two lists and stores the devices that are only in Nessus. The script then loops through this list of devices and performs a scan on each one using Nessus. The script can be scheduled to run at a specific time in your environment using cron or other scheduling tools.